LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/onerroreventhandler-expected.txt



FAIL onerror + ErrorEvent + Window assert_equals: expected 5 but got 1
FAIL onerror + !ErrorEvent + Window assert_equals: expected 5 but got 1
PASS onerror + Document
